act_acq/CMakeLists.txt - Created build script for new act_acq act_acq/acq_net.c - Several minor bug fixes act_acq/acq_store.c - Several minor changes and bug fixes act_acq/act_acq.c - Finally finished implementing act_acq/ccd_cntrl.h - Removed old TODO comments act_acq/ccd_cntrl.c   - (MAJOR) Implemented integration repetitions - TEST!!!   - (MINOR) Removed code that stores approximate integration start time at start of integration - merlin driver now reports full date and time in image parameters structure, so th$   - (TRIVIAL) Removed all date and time code from cntrl structure act_acq/expose_dialog.* - Minor changes act_acq/imgdisp.c - Trivial changes act_control/net_dataccd.c - Updated for new IPC (no supported window/prebin modes handshake in ccdcap messages) act_dti/dti_net.c - Updated for new IPC act_dti/telmove.c - Updated for new IPC act_environ/act_environ.c - Updated for new IPC act_environ/env_weather.c - Updated for new IPC act_pmpthot/act_pmtphot.c - Updated for new IPC act_sched/act_sched.c - Updated for new IPC

+++ softw/act_dti/dti_net.c
@@ -409,7 +409,7 @@ DtiNet *dti_net_new (const gchar *host, const gchar *port)
 
 gint dti_net_send(DtiNet *dti_net, DtiMsg *msg)
 {
-  unsigned int num_bytes;
+  gsize num_bytes;
   int status;
   GError *error = NULL;
   status = g_io_channel_write_chars (dti_net->net_chan, (gchar *)(&msg->msg), sizeof(struct act_msg), &num_bytes, &error);
@@ -474,7 +474,7 @@ static gboolean net_read_ready(GIOChannel *net_chan, GIOCondition cond, gpointer
   (void) dti_net;
 //   DtiNet *objs = DTI_NET(dti_net);
   struct act_msg msgbuf;
-  unsigned int num_bytes;
+  gsize num_bytes;
   int status;
   GError *error = NULL;
   status = g_io_channel_read_chars (net_chan, (gchar *)&msgbuf, sizeof(msgbuf), &num_bytes, &error);
